+++ src/v1/HttpStatusCode/HttpStatusCode.ts
@@ -31,7 +31,14 @@
 // ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 //
-import { Branded } from "@ganbarodigital/ts-lib-value-objects/lib/V1";
+
+// copied from ts-lib-value-objects, so that we can break the
+// circular dependency
+interface Branding<BrandingT extends string> {
+    BrandingT: BrandingT;
+}
+
+type Branded<T, BrandingT extends string> = T & Branding<BrandingT>;
 
 /**
  * represents any HTTP status code in the range `100` to `599` inclusive