+++ CHANGELOG.md
@@ -21,6 +21,16 @@ For each release, changes are grouped under these headings:
 
 The following changes have been completed, and will be included in the next tagged release.
 
+## v0.1.3
+
+### Fixes
+
+* Introduced `internal.ts` to resolve circular dependencies within the package.
+
+### Dependencies
+
+* Remove dependenciy on `@ganbarodigital/ts-lib-packagename` (causes circular dependency hell)
+
 ## v0.1.2
 
 Released Thursday, 20th February 2020.